refactor: streamline product detail page by modularizing components
- Removed unused icons and components to simplify the codebase.
- Introduced new components: ProductHeader, ProductSummary, ProductHighlights, ProductPreview, ProductTags, and DesignToolIcons for better organization and readability.
- Enhanced layout and structure of the product detail page for improved user experience.
